requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
django-filter==24.2
python-dotenv==1.0.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
Pillow==10.4.0
boto3==1.34.162
django-storages==1.14.4
django-modeltranslation==0.18.11
drf-spectacular==0.27.2
